Clean-up / fix tests
