+++ src/UnitTests/SessionExtensionsTests.cs
@@ -21,50 +21,6 @@ namespace UnitTests
         }
 
         [Test]
-        public void SetNode_Should_Throw_Exception_If_NodeIdAttribute_Is_Not_Found()
-        {
-            var session = Substitute.For<ISession>();
-
-            var exception = Assert.Throws<InvalidOperationException>(
-                () => session.SetNode(new NonEntity()));
-
-            Assert.AreEqual("NodeIdAttribute not specified or the Node Id is null", exception.Message);
-        }
-
-        [Test]
-        public void SetNode_Should_Throw_Exception_If_Node_Id_Value_Is_Null()
-        {
-            var session = Substitute.For<ISession>();
-
-            var exception = Assert.Throws<InvalidOperationException>(
-                () => session.SetNode(new EntityWithNullableId()));
-
-            Assert.AreEqual("NodeIdAttribute not specified or the Node Id is null", exception.Message);
-        }
-
-        [Test]
-        public void SetNode_Tx_Should_Throw_Exception_If_NodeIdAttribute_Is_Not_Found()
-        {
-            var transaction = Substitute.For<ITransaction>();
-
-            var exception = Assert.Throws<InvalidOperationException>(
-                () => transaction.SetNode(new NonEntity()));
-
-            Assert.AreEqual("NodeIdAttribute not specified or the Node Id is null", exception.Message);
-        }
-
-        [Test]
-        public void SetNode_Tx_Should_Throw_Exception_If_Node_Id_Value_Is_Null()
-        {
-            var transaction = Substitute.For<ITransaction>();
-
-            var exception = Assert.Throws<InvalidOperationException>(
-                () => transaction.SetNode(new EntityWithNullableId()));
-
-            Assert.AreEqual("NodeIdAttribute not specified or the Node Id is null", exception.Message);
-        }
-
-        [Test]
         public void SetNodeAsync_Should_Throw_Exception_If_NodeIdAttribute_Is_Not_Found()
         {
             var session = Substitute.For<IAsyncSession>();
